+++ tests/tests.py
@@ -161,6 +161,7 @@ class TestSomething(unittest.TestCase) :
 
     def test_color_themes(self) :
         originalThemes = {
+            "batty",
             "light",
             "light-colorblind",
             "light-high-contrast",
@@ -169,7 +170,9 @@ class TestSomething(unittest.TestCase) :
             "dark-colorblind",
             "dark-dimmed",
             "dark-high-contrast",
-            "dark-tritanopia"
+            "dark-tritanopia",
+            "halloween",
+            "halloween-light"
         }
         for theme in originalThemes :
             self._colorValidation(colorMapping[theme])
@@ -305,7 +308,7 @@ class TestSomething(unittest.TestCase) :
         stats = NoQueries(True, False, 100, set(), "FavoriteRepo")
         #categories = ["general", "repositories", "languages", "contributions"]
         categories = categoryOrder[:]
-        colors = copy.deepcopy(colorMapping["dark"])
+        colors = copy.deepcopy(colorMapping["batty"])
         #colors["title-icon"] = "pumpkin"
         svgGen = StatsImageGenerator(
             stats,